+++ src/redturtle/prenotazioni/adapters/booker.py
@@ -11,6 +11,7 @@ from zope.interface import implementer
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni.prenotazione_event import MovedPrenotazione
 from redturtle.prenotazioni.utilities.dateutils import exceedes_date_limit
+from redturtle.prenotazioni.content.prenotazione import VACATION_TYPE
 from zope.event import notify
 from redturtle.prenotazioni import datetime_with_tz
 from six.moves.urllib.parse import parse_qs
@@ -247,11 +248,9 @@ class Booker(object):
         for period in ("morning", "afternoon"):
             free_slots = self.prenotazioni.get_free_slots(start, period)
             gate_free_slots = free_slots.get(gate, [])
-            [
-                slots.append(vacation_slot.intersect(slot))
-                for slot in gate_free_slots
-                if vacation_slot.overlaps(slot)
-            ]
+            for slot in gate_free_slots:
+                if vacation_slot.overlaps(slot):
+                    slots.append(vacation_slot.intersect(slot))
 
         start_date = DateTime(start.strftime("%Y/%m/%d"))
         for slot in slots:
@@ -261,4 +260,6 @@ class Booker(object):
             # duration = float(len(slot)) / 60
             slot_data = {k: v for k, v in data.items() if k != "gate"}
             slot_data["booking_date"] = datetime_with_tz(booking_date)
+            slot_data["booking_type"] = VACATION_TYPE
             self.create(slot_data, duration=duration, force_gate=gate)
+        return len(slots)