+++ src/redturtle/prenotazioni/config.py
@@ -14,3 +14,4 @@ REQUIRABLE_AND_VISIBLE_FIELDS = [
     "company",
     "description",
 ]
+DEFAULT_VISIBLE_BOOKING_FIELDS = ["email", "phone", "description"]