+++ src/redturtle/prenotazioni/restapi/serializers/adapters/prenotazione.py
@@ -51,6 +51,7 @@ class PrenotazioneSerializer:
                 status["review_state"], context=self.request
             ),
             "booking_type": self.prenotazione.booking_type,
+            "vacation": self.prenotazione.isVacation(),
             "booking_code": self.prenotazione.getBookingCode(),
             "cosa_serve": useful_docs,
         }
@@ -89,4 +90,5 @@ class PrenotazioneSearchableItemSerializer:
             "phone": self.prenotazione.phone,
             "staff_notes": self.prenotazione.staff_notes,
             "company": self.prenotazione.company,
+            "vacation": self.prenotazione.isVacation(),
         }