+++ CHANGES.rst
@@ -7,6 +7,11 @@ Changelog
 
 - Tabs/fields reordering for the booking folder.
   [daniele]
+- fix date in @@download
+  [mamico]
+
+- skip email to manager on block/vacation creation
+  [mamico]
 
 - Manage timezone in booking dates. (upgrade step)
   [cekk]
@@ -37,6 +42,9 @@ Changelog
 
 - Rename routes:
   months-slots => available-slots
+  prenotazione-schema => booking-schema
+  @@download_reservation => @@download/bookings.xlsx
+  [cekk] [mamico]  
   
 
 2.0.0.dev3 (2023-07-20)